#6650. Replaced return type of GetClockTicks & ClockTicksPerSecond from double to uint64.
The main reason for it is less overhead.
When this timer gets used for measuring time of R_AddSingleModel call, uint64->double conversion adds +33% to GetClockTicks call.
And this overhead makes no sense, because the timer is integer, and high-performance users also expect it to be integer.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@11076 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ sys/win32/win_cpu.cpp
@@ -35,7 +35,7 @@ Project: The Dark Mod (http://www.thedarkmod.com/)
 Sys_GetClockTicks
 ================
 */
-double Sys_GetClockTicks( void ) {
+uint64 Sys_GetClockTicks( void ) {
 #if 0
 
 	LARGE_INTEGER li;
@@ -50,7 +50,7 @@ double Sys_GetClockTicks( void ) {
 	__cpuid( values, 0 );*/
 	// greebo: Use the intrinsic provided by the VC++ compiler in x64
 	unsigned __int64 ticks = __rdtsc();
-	return static_cast<double>( ticks );
+	return ticks;
 
 #endif
 }
@@ -60,8 +60,8 @@ double Sys_GetClockTicks( void ) {
 Sys_ClockTicksPerSecond
 ================
 */
-double Sys_ClockTicksPerSecond( void ) {
-	static double ticks = 0;
+uint64 Sys_ClockTicksPerSecond( void ) {
+	static uint64 ticks = 0;
 #if 0
 
 	if ( !ticks ) {
